move slice types to interfaces file
