+++ client/src/features/sceneControls/mathItems/mathItemsSlice/index.ts
@@ -1,9 +1,10 @@
 import * as select from "./selectors";
-import type { MathItemsState } from "./mathItems.slice";
+import type { Subtree, MathItemsState } from "./interfaces";
 import mathItemsSlice, { actions, reducer } from "./mathItems.slice";
 
 export { select, actions, reducer };
 export * from "./hooks";
 
+export type { Subtree };
 export type { MathItemsState };
 export default mathItemsSlice;
