+++ client/src/features/sceneControls/mathItems/mathItemsSlice/mathItems.slice.ts
@@ -10,23 +10,7 @@ import {
   syncItemsToMathScope,
   removeItemsFromMathScope,
 } from "./syncMathScope";
-
-interface MathItemsState {
-  items: {
-    [id: string]: MathItem;
-  };
-  order: Record<string, string[]>;
-  activeItemId: string | undefined;
-  /**
-   * We need to sync the expressions in MathItem.properties with the MathScope.
-   * Putting mathScope in the redux store is a very convenient way to do this
-   * since actions are the centralized place for editing MathItem properties.
-   *
-   * This is a nonserializable value, so comes with some caveats. See
-   * https://redux-toolkit.js.org/usage/usage-guide#working-with-non-serializable-data
-   */
-  mathScope: () => MathScope;
-}
+import type { MathItemsState } from "./interfaces";
 
 const makeMathScope = () => new MathScope({ parse: latexParser.parse });
